+++ test/test_helper.exs
@@ -1,17 +1,17 @@
 ExUnit.start()
 
 Volt.Test.ExUnit.install(
-  root: "test/astral/islands/fixtures",
-  include: ["*.browser.test.ts"],
+  root: "test/astral/islands",
+  include: ["browser/*.test.ts"],
   browser: true,
   bundle: [
     node_modules: "node_modules",
     plugins: [
       Astral.Islands.RuntimePlugin,
+      Astral.Islands.SolidPlugin,
       Volt.Plugin.Vue,
       Volt.Plugin.Svelte,
-      Volt.Plugin.React,
-      Volt.Plugin.Solid
+      Volt.Plugin.React
     ]
   ]
 )